add some rust test case

+++ wasmtest/ccntmracts-cplus/test_callwasmccntmract.cpp
@@ -16,8 +16,8 @@ class hello: public ccntmract {
 	public:
 	using ccntmract::ccntmract;
 
-	int64_t call_wasm_ccntmract(int64_t a, int64_t b, test_conext &tc) {
-		int64_t res;
+	int128_t call_wasm_ccntmract(int128_t a, int128_t b, test_conext &tc) {
+		int128_t res;
 		address test_add = tc.addrmap["test_add.wasm"];
 		auto args = pack(string("add"), a, b);
 		call_ccntmract(test_add, args, res);
